+++ src/formatters/phone-number.ts
@@ -1,4 +1,4 @@
-import { COUNTRY_CODE } from "../constants/phone-number";
+import { COUNTRY_CODE } from "../constants";
 
 /**
  * Formats a Tunisian phone number
