+++ src/index.ts
@@ -1,28 +1,4 @@
-// Export validators
-export { validateCIN } from "./validators/cin";
-export { validatePhoneNumber, getCarrierInfo } from "./validators/phone-number";
-export { validatePostalCode } from "./validators/postal-code";
-export { validateTaxID } from "./validators/tax-id";
-export { validateRIB, getBankFromRIB } from "./validators/bank";
-
-// Export formatters
-export { formatPhoneNumber } from "./formatters/phone-number";
-export { formatCurrency } from "./formatters/currency";
-export { formatDate } from "./formatters/date";
-
-// Export constants
-export { BANKS } from "./constants/banks";
-export { CARRIERS, COUNTRY_CODE } from "./constants/phone-number";
-export { GOVERNORATES, POSTAL_CODES_MAP } from "./constants/regions";
-
-// Export types
-export type {
-  CIN,
-  PhoneNumber,
-  PostalCode,
-  TaxID,
-  RIB,
-  Bank,
-  Carrier,
-  CurrencyFormatOptions,
-} from "./types";
+export * from "./validators";
+export * from "./formatters";
+export * from "./constants";
+export * from "./types";
